Update parsing and regression notebook
- parsing_synthetic_data notebook is now more general and add a delta_E
column by approximating the reference energy by using the simulation
with the highest hyperparameters.
- toy_regression notebook now concatenates several datasetsand the prediction
are made on delta_E instead of total_energy
- in the toy_regression notebook, the feature augmentation has been
reduced. For linear regression, a polynomial degree of 2 seems to work
far better than 3 (maybe because the data was not scaled before the
computation of polynomial features).
For random forest, the model seems to work pretty well without feature
augmentation.